fixed bugs and clean up. New crash need investigation

+++ Core/Source/Sprite.cpp
@@ -28,8 +28,8 @@ void SDL_Sprite::initialize() {
 }
 
 void SDL_Sprite::setDestPosition(float x, float y) {
-    destRect.x = x;
-    destRect.y = y;
+    destRect.x = (int)x;
+    destRect.y = (int)y;
 }
 
 SDL_Rect* SDL_Sprite::getDestRect() {
@@ -37,10 +37,10 @@ SDL_Rect* SDL_Sprite::getDestRect() {
 }
 
 void SDL_Sprite::setSrcPosition(float x, float y, float w, float h) {
-    srcRect.x = x;
-    srcRect.y = y;
-    srcRect.w = w;
-    srcRect.h = h;
+    srcRect.x = (int)x;
+    srcRect.y = (int)y;
+    srcRect.w = (int)w;
+    srcRect.h = (int)h;
 }
 
 SDL_Rect* SDL_Sprite::getSrcRect() {
@@ -54,8 +54,8 @@ void SDL_Sprite::setRotation(double angle) {
 void SDL_Sprite::setScale(float scaleX, float scaleY) {
     this->scaleX = scaleX;
     this->scaleY = scaleY;
-    destRect.w = srcRect.w * scaleX;
-    destRect.h = srcRect.h * scaleY;
+    destRect.w = srcRect.w * (int)scaleX;
+    destRect.h = srcRect.h * (int)scaleY;
     center = {destRect.w / 2.0f, destRect.h / 2.0f};
 }
